on invite accepted go to main game page fix

+++ TriviaPatente/app/src/main/java/com/ted_developers/triviapatente/models/game/Invite.java
@@ -13,7 +13,7 @@ public class Invite extends Base {
     @SerializedName("sender_id")
     public Long sender_id;
     @SerializedName("game_id")
-    public Integer game_id;
+    public Long game_id;
     @SerializedName("sender_name")
     public String sender_name;
     @SerializedName("sender_surname")
